refactor: extract astra-pipeline crate with routing types
Phase 1 of runtime crate splitting:

- Create astra-pipeline crate with core routing types:
  - TaskType (8-way task classification)
  - DomainHint (7 domain categories)
  - ToolFilter (tool selection strategy)
  - CalibrationAxis (calibration targeting)
  - domain_hint_to_label helper

- Update runtime to re-export from astra-pipeline
- Remove duplicate CalibrationAxis from evolution/types.rs

This enables other crates to depend on routing types without
pulling in the full runtime crate.

LOC impact: ~80 lines moved to astra-pipeline
