+++ rust/crates/runtime/Cargo.toml
@@ -31,6 +31,7 @@ path = "src/main.rs"
 dirs = "6.0.0"
 fastrand = "2"
 astra-core.workspace = true
+astra-pipeline.workspace = true
 astra-services.workspace = true
 astra-tools.workspace = true
 astra-thin-client.workspace = true
